% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.data.frame.gradient}
\alias{as.data.frame.gradient}
\title{Convert gradient to data frame}
\usage{
\method{as.data.frame}{gradient}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{see data frame}

\item{row.names}{see data frame}

\item{optional}{see data frame}

\item{...}{see data frame}
}
\value{
data frame
}
\description{
Convert gradient to data frame
}

