% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.data.frame.hessian}
\alias{as.data.frame.hessian}
\title{Convert hessian to data frame}
\usage{
\method{as.data.frame}{hessian}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{see data frame}

\item{row.names}{see data frame}

\item{optional}{see data frame}

\item{...}{see data frame}
}
\value{
data frame
}
\description{
Convert hessian to data frame
}

