% Generated by roxygen2 (4.0.1): do not edit by hand
\name{chunkNames}
\alias{chunkNames}
\title{Get chunk names from a knitr document}
\usage{
chunkNames(doc)
}
\arguments{
\item{doc}{character string of a \code{knitr} document}
}
\value{
character string vector with names of chunks (basically
... it's good enough anyways)
}
\description{
Get chunk names from a knitr document
}

