% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparing_classes.R
\name{comparing_classes}
\alias{comparing_classes}
\title{Function for comparing the classes of tow data sets for if I am having difficulty joining them.}
\usage{
comparing_classes(df1, df2)
}
\arguments{
\item{df1}{an object of class "data.frame".}

\item{df2}{an object of class "data.frame".}
}
\value{
data.frame object
}
\description{
Function for comparing the classes of tow data sets for if I am having difficulty joining them.
}
\examples{
comparing_classes(df1,df2)
comparing_classes(
  df1 = cars,
  df2 = mtcars \%>\% select(mpg,cyl)
  )
}
