% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dontStopLmer}
\alias{dontStopLmer}
\title{Don't stop lmer}
\usage{
dontStopLmer(cntrl = lmerControl)
}
\arguments{
\item{cntrl}{either \code{lmerControl} or \code{glmerControl}}
}
\value{
list for \code{control} argument in \code{lmer} or
\code{glmer}
}
\description{
Don't stop lmer
}

