% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_loop.R
\name{double_loop}
\alias{double_loop}
\title{Function for looping through data.frame and apply operations to specific values.}
\usage{
double_loop()
}
\description{
Function for looping through data.frame and apply operations to specific values.
}
