% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_info.R
\name{get_file_info}
\alias{get_file_info}
\title{Function for getting file information}
\usage{
get_file_info(path)
}
\arguments{
\item{path}{A class of "string" that is the file path I want info for.}
}
\value{
data table
}
\description{
Function for getting file information
}
\examples{
get_file_info(getwd())
get_file_info(file.path(getwd(),"output"))
}
