% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gitLines}
\alias{gitLines}
\title{Count number of insertions}
\usage{
gitLines(dir = ".", args = NULL)
}
\arguments{
\item{dir}{git repo}

\item{args}{additional arguments for \code{git log}}
}
\value{
total number of additions
}
\description{
Count number of insertions
}

