% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hessian}
\alias{hessian}
\title{Hessians}
\usage{
hessian(expr, variables, hessianOnly = TRUE)
}
\arguments{
\item{expr}{an expression}

\item{variables}{variables}

\item{hessianOnly}{well...}
}
\value{
hessian object or list of a gradient and hessian object
}
\description{
Hessians
}

