% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadData}
\alias{loadData}
\title{Run all lines in a document containing \code{load} and \code{Rda}}
\usage{
loadData(doc)
}
\arguments{
\item{doc}{character string of a document name}
}
\value{
run lines in \code{.GlobalEnv}
}
\description{
Run all lines in a document containing \code{load} and \code{Rda}
}

