% Generated by roxygen2 (4.0.1): do not edit by hand
\name{purlChunks}
\alias{purlChunks}
\title{Purl a subset of knitr chunks}
\usage{
purlChunks(doc, chunkNames, filePrefix = ".")
}
\arguments{
\item{doc}{character string of a \code{knitr} document}

\item{chunkNames}{character string vector of chunk names}

\item{filePrefix}{character string to \code{\link{paste}} in front
of \code{doc}}
}
\value{
writes new \code{R} file
}
\description{
Purl a subset of knitr chunks
}
\note{
Every chunk must be named for consistent results
}

