% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spss.R
\name{read_spss}
\alias{read_spss}
\title{Function for reading in and pre-processing SPSS data}
\usage{
read_spss(path, skim = TRUE)
}
\arguments{
\item{path}{A class of "string" that is the file path to your data set.}

\item{skim}{A class of "logical" that indicates if I want to print a summary of my data from the skimr package.}
}
\value{
table
}
\description{
Function for reading in and pre-processing SPSS data
}
\examples{
data = read_spss(path = "data/2024 Reports 03.19.24.sav")
data = read_spss(path = "data/2024 Reports 03.19.24.sav", skim = TRUE)
data = read_spss(path = "data/2024 Reports 03.19.24.sav", skim = FALSE)
}
