% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rmat}
\alias{rmat}
\title{Random matrix}
\usage{
rmat(nrow = 1, ncol = 1, rfunc = rnorm, ...)
}
\arguments{
\item{rfunc}{an \code{r} function (e.g. \code{\link{rnorm}})}

\item{nrow,ncol}{number of rows and columns}

\item{...}{parameters to \code{rfunc}}
}
\value{
a random matrix
}
\description{
Random matrix
}

