% Generated by roxygen2 (4.0.1): do not edit by hand
\name{runChunks}
\alias{runChunks}
\title{Run a subset of knitr chunks}
\usage{
runChunks(doc, chunkNames)
}
\arguments{
\item{doc}{character string of a \code{knitr} document}

\item{chunkNames}{character string vector of chunk names}
}
\value{
run chunks in \code{.GlobalEnv}
}
\description{
Run a subset of knitr chunks
}
\note{
Every chunk must be named for consistent results
}

