% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sourceVersion}
\alias{sourceVersion}
\title{Version of package in local source code}
\usage{
sourceVersion(pkg)
}
\arguments{
\item{pkg}{package name as character string}
}
\description{
Meant for use with makefiles.  Must be in the directory
with the package source.
}

