% Generated by roxygen2 (4.0.1): do not edit by hand
\name{UniformInt}
\alias{UniformInt}
\alias{dunifInt}
\alias{runifInt}
\title{The Uniform Distribution over the Integers}
\usage{
runifInt(n, min = 0, max = 1)

dunifInt(x, min = 0, max = 1, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{min}{lower limit (if not integer, it is rounded)}

\item{max}{upper limit (if not integer, it is rounded)}

\item{n}{number of observations}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
The Uniform Distribution over the Integers
}

